+++ R/MAPAssignActivity.R
@@ -93,7 +93,8 @@ MAPAssignActivity <-
            drop.low.mut.samples       = TRUE,
            use.sig.presence.test      = FALSE,
            sig.pres.test.nbinom.size  = NULL,
-           sig.pres.test.p.thresh     = 0.05) {
+           sig.pres.test.p.thresh     = 0.05,
+           sig.pres.test.q.thresh     = NULL) {
     if (drop.low.mut.samples) {
       spectra <- DropLowMutationSamples(spectra)
     } else {
@@ -121,7 +122,8 @@ MAPAssignActivity <-
         drop.low.mut.samples        = drop.low.mut.samples, 
         use.sig.presence.test       = use.sig.presence.test,
         sig.pres.test.nbinom.size   = sig.pres.test.nbinom.size,
-        sig.pres.test.p.thresh      = sig.pres.test.p.thresh)
+        sig.pres.test.p.thresh      = sig.pres.test.p.thresh,
+        sig.pres.test.q.thresh      = sig.pres.test.q.thresh)
 
       return(retval1)
     }