+++ R/RunMAPOneSample.R
@@ -27,7 +27,8 @@ RunMAPOnOneSample <-
            drop.low.mut.samples        = TRUE,
            use.sig.presence.test       = FALSE,
            sig.pres.test.nbinom.size   = NULL,
-           sig.pres.test.p.thresh      = 0.05) {
+           sig.pres.test.p.thresh      = 0.05,
+           sig.pres.test.q.thresh      = NULL) {
     
     if (drop.low.mut.samples) {
       spect <- DropLowMutationSamples(spect)
@@ -55,7 +56,8 @@ RunMAPOnOneSample <-
       drop.low.mut.samples       = drop.low.mut.samples, 
       use.sig.presence.test      = use.sig.presence.test,
       sig.pres.test.nbinom.size  = sig.pres.test.nbinom.size,
-      sig.pres.test.p.thresh     = sig.pres.test.p.thresh)
+      sig.pres.test.p.thresh     = sig.pres.test.p.thresh,
+      sig.pres.test.q.thresh     = sig.pres.test.q.thresh)
     
     if (!is.null(retval$error.messages)) {
       return(retval)