+++ R/SigPresenceAssignActivity.R
@@ -79,7 +79,8 @@ SigPresenceAssignActivity <-
            seed                      = NULL,
            drop.low.mut.samples      = TRUE,
            sig.pres.test.nbinom.size = NULL,
-           sig.pres.test.p.thresh    = 0.05) {
+           sig.pres.test.p.thresh    = 0.05,
+           sig.pres.test.q.thresh    = NULL) {
     
     retval <- 
       SparseAssignActivity(spectra                   = spectra,
@@ -96,6 +97,7 @@ SigPresenceAssignActivity <-
                            drop.low.mut.samples      = drop.low.mut.samples,
                            use.sig.presence.test     = TRUE,
                            sig.pres.test.nbinom.size = sig.pres.test.nbinom.size,
-                           sig.pres.test.p.thresh    = sig.pres.test.p.thresh)
+                           sig.pres.test.p.thresh    = sig.pres.test.p.thresh,
+                           sig.pres.test.q.thresh    = sig.pres.test.q.thresh)
     return(retval)
   }