+++ R/SparseAssignActivity.R
@@ -69,7 +69,8 @@ SparseAssignActivity <-
            drop.low.mut.samples      = TRUE,
            use.sig.presence.test     = FALSE,
            sig.pres.test.nbinom.size = NULL,
-           sig.pres.test.p.thresh    = 0.05) {
+           sig.pres.test.p.thresh    = 0.05,
+           sig.pres.test.q.thresh    = NULL) {
     
     retval <- 
       MAPAssignActivity(spectra                   = spectra,
@@ -87,7 +88,8 @@ SparseAssignActivity <-
                         drop.low.mut.samples      = drop.low.mut.samples,
                         use.sig.presence.test     = use.sig.presence.test,
                         sig.pres.test.nbinom.size = sig.pres.test.nbinom.size,
-                        sig.pres.test.p.thresh    = sig.pres.test.p.thresh)
+                        sig.pres.test.p.thresh    = sig.pres.test.p.thresh,
+                        sig.pres.test.q.thresh    = sig.pres.test.q.thresh)
     return(retval)
   }